requirements.txt
asgiref==3.2.7
Django==3.0
django-ace==1.0.7
djangorestframework==3.11.0
djangorestframework-stubs==1.2.0
pytz==2020.1
sqlparse==0.3.1
